feat: Report progress of Db compaction

+++ include/je2be/tobe/db/db.hpp
@@ -6,7 +6,7 @@ class Db : public DbInterface {
 public:
   Db(std::string const &) = delete;
   Db(std::wstring const &) = delete;
-  explicit Db(std::filesystem::path const &dir) : fDb(nullptr), fValid(false) {
+  explicit Db(std::filesystem::path const &dir) : fDb(nullptr) {
     using namespace leveldb;
 
     DB *db;
@@ -18,29 +18,34 @@ public:
       return;
     }
     fDb = db;
-    fValid = true;
   }
 
   ~Db() {
-    if (fValid) {
+    if (fDb) {
       delete fDb;
     }
   }
 
-  bool valid() const override { return fValid; }
+  bool valid() const override { return fDb != nullptr; }
 
   void put(std::string const &key, leveldb::Slice const &value) override {
-    assert(fValid);
-    fDb->Put(fWriteOptions, key, value);
+    assert(fDb);
+    if (fDb) {
+      fDb->Put(fWriteOptions, key, value);
+    }
   }
 
   void write(leveldb::WriteBatch &batch) {
     assert(fValid);
-    fDb->Write(leveldb::WriteOptions{}, &batch);
+    if (fDb) {
+      fDb->Write(leveldb::WriteOptions{}, &batch);
+    }
   }
 
   std::optional<std::string> get(std::string const &key) {
-    assert(fValid);
+    if (!fDb) {
+      return std::nullopt;
+    }
     leveldb::ReadOptions o;
     std::string v;
     leveldb::Status st = fDb->Get(o, key, &v);
@@ -52,16 +57,37 @@ public:
   }
 
   void del(std::string const &key) override {
-    assert(fValid);
-    fDb->Delete(leveldb::WriteOptions{}, key);
+    if (fDb) {
+      fDb->Delete(leveldb::WriteOptions{}, key);
+    }
+  }
+
+  bool close(std::optional<std::function<void(double progress)>> progress = std::nullopt) override {
+    if (!fDb) {
+      return false;
+    }
+    if (progress) {
+      (*progress)(0.0);
+    }
+    if (fDb) {
+      delete fDb;
+      fDb = nullptr;
+    }
+    if (progress) {
+      (*progress)(1.0);
+    }
+    return true;
   }
 
   void abandon() override {
+    if (fDb) {
+      delete fDb;
+      fDb = nullptr;
+    }
   }
 
 private:
   leveldb::DB *fDb;
-  bool fValid;
   leveldb::WriteOptions fWriteOptions;
 };
